move mining errors to a lower crate and use it for rpc core (which targets wasm)
