+++ mining/Cargo.toml
@@ -13,6 +13,7 @@ kaspa-muhash.workspace = true
 kaspa-consensus-core.workspace = true
 kaspa-txscript.workspace = true
 kaspa-core.workspace = true
+kaspa-mining-errors.workspace = true
 kaspa-consensusmanager.workspace = true
 thiserror.workspace = true
 serde.workspace = true
