+++ mining/errors/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "kaspa-mining-errors"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+thiserror.workspace = true
+kaspa-consensus-core.workspace = true
