+++ mining/errors/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod block_template;
+pub mod manager;
+pub mod mempool;
